Merged PR 631: [feature] properly filter log records when they're modified by a logging.filter
When you modify a LogRecord inside a filter, whether it should be emitted or not based on the current level is determined by the handler before it applies the filter in the base class. This just reverses the order of the evaluation to make sure we evaluate the filter before the level. It also includes pieces getting a named logger rather than the implicit root logger.